MH-13191 Improve performance of event deletion
When deleting events from the scheduler service, some unnecessary
queries related to properties are executed:

- Since we know, that we are deleting the whole episode, we do not have
to check whether there are snapshots for the properties to be deleted. I
am not sure anyway, why this is necessary at all.

- Since we know, that we take care of deleting all properties related to
the episode, we do not have to remove orphaned properties (because we are
not producing any).

It turned out to be quiet hard to control the behavior of the code
executed in AbstractADeleteQuery.runQueries() and I am not happy with
the flag attached to DeleteQueryContribution. However, I could not find a
better solution in the available time.

All in all, these changes seem to improve event deletion by about 50 or
60%, according to our measurements.

+++ modules/asset-manager-impl/src/main/java/org/opencastproject/assetmanager/impl/query/AbstractADeleteQuery.java
@@ -93,6 +93,16 @@ public abstract class AbstractADeleteQuery implements ADeleteQuery, DeleteQueryC
     };
   }
 
+  @Override public ADeleteQuery willRemoveWholeMediaPackage(boolean willRemoveWholeMediaPackage) {
+    return new AbstractADeleteQuery(am, owner) {
+      @Override
+      public DeleteQueryContribution contributeDelete(String owner) {
+        final DeleteQueryContribution cParent = AbstractADeleteQuery.this.contributeDelete(owner);
+        return DeleteQueryContribution.mk(cParent).willRemoveWholeMediaPackage(willRemoveWholeMediaPackage);
+      }
+    };
+  }
+
   public long run(DeleteSnapshotHandler deleteSnapshotHandler) {
     // run query and map the result to records
     final long startTime = System.nanoTime();
@@ -192,7 +202,9 @@ HAVING v = (SELECT count(*)
       am.getDb().logDelete(formatQueryName(c.name, "main"), qMain);
       final long deletedItems = qMain.execute();
       // delete orphaned properties
-      deleteOrphanedProperties();
+      if (!c.willRemoveWholeMediaPackage) {
+        deleteOrphanedProperties();
+      }
       // <BLOCK>
       // TODO Bad solution. Yields all media package IDs which can easily be thousands
       // TODO The above SQL solution does not work with H2 so I suspect the query is not 100% clean
@@ -220,7 +232,7 @@ HAVING v = (SELECT count(*)
       final BooleanExpression where;
       {
         final BooleanExpression w = c.where.apply(Q_PROPERTY);
-        if (w != null) {
+        if (w != null && !c.willRemoveWholeMediaPackage) {
           /* The original sub query used an "ON" clause to filter the join by mediapackage id [1].
              Unfortunately Eclipse link drops this clause completely when transforming the query
              into SQL. It creates a cross join instead of the inner join, which is perfectly legal
@@ -248,7 +260,7 @@ HAVING v = (SELECT count(*)
                           .distinct()
                           .list(Q_PROPERTY.mediaPackageId));
         } else {
-          where = null;
+          where = w;
         }
       }
       final JPADeleteClause qProperties = jpa.delete(from).where(Expressions.allOf(c.targetPredicate.orNull(), where));
